Matrix class skeleton: Constructors and Destructor
This is the very first version of the matrix class. Private data
members: matrix (the actual 2-D matrix), rows and cols (denoting the
number of rows and columns in the 2-D matrix). A parameterized
constructor, a copy constructor and a destructor have been added. The
parameterized constructor takes 3 arguments : the number of rows,
columns and the initial value (all entries of the matrix will be
initialized with this value).