Added methods for display and scalar multipy
Method to multiply a matrix with a scalar has been added alond with a
function to display a matrix. The display function is useful for testing
and debug purposes. A few minor corrections were made to the existing
methods for basic arithmetic. All the code related to testing resides in
main.cpp.