LU Decomposition
Factorizes the matrix A into Lower Triangular Matrix L
,Upper Triagular Matrix U and a Permutation Matrix P
which is required to satisfy PA = LU. For the permu-
tation matrix the function doesn't directly take n-by-n
input but instead it takes a vector<int> which contains
the row indices of 1 as permutation matrix consists of
only 1 and 0.

Suppose after executing the function a 4-by-1 P is given
by '0 2 3 1' then the corresponding 4-by-4 permutation
matrix can generated by setting Perm(0, 0) , Perm(1, 2),
Perm(2, 3) and Perm(3, 1) to 1.